Allow users to save a simulation that is not part of a batch.
There are now three types of simulations: simulations that have been
ran, batched, or saved. All simulations in a batch will be run
as a an entire set. They will then appear in the "ran" list.

Alternatively, a user can just "save" the simulation without putting
it in a batch to run it individually later.

All simulations must have unique names, regardless of its state.